initial commit for this enormous PR
