+++ contextio/lib/v2_0/helpers.py
@@ -1,0 +1,127 @@
+from datetime import datetime
+import logging
+import re
+
+class ArgumentError(Exception):
+    """Class to handle bad arguments."""
+    def __init__(self, message):
+        self.message = message
+
+    def __str__(self):
+        return self.message
+
+def to_underscore(name):
+    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
+    return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
+
+
+def uncamelize(d):
+    drop = []
+
+    for k, v in list(d.items()):
+        u = to_underscore(k)
+        if u != k and u not in d:
+            d[u] = v
+            drop.append(k)
+
+    for k in drop:
+        del d[k]
+
+    return d
+
+
+def as_bool(v):
+    if v == 0 or v is False:
+        return False
+    return True
+
+
+def as_datetime(v):
+    if isinstance(v, int):
+        return datetime.fromtimestamp(v)
+
+
+def process_person_info(parent, person_info, addresses):
+    try:
+        from contextIO2 import Contact
+    except ImportError:
+        from .__init__ import Contact
+    contacts = {}
+    to_addrs = []
+    to_contacts = []
+    from_addr = None
+    from_contact = None
+
+    if 'to' in addresses:
+        for info in addresses['to']:
+            person_info[info.get('email')].setdefault('name', info.get('name'))
+            to_addrs.append(info.get('email'))
+
+    info = addresses['from']
+    person_info[info.get('email')].setdefault('name', info.get('name'))
+    from_addr = info.get('email')
+
+    for addr, d in list(person_info.items()):
+        info = {
+            'email': addr,
+            'thumbnail': d.get('thumbnail'),
+            'name': d.get('name')
+        }
+        c = Contact(parent, info)
+        contacts.setdefault(addr, c)
+
+        if addr in to_addrs:
+            to_contacts.append(c)
+            to_addrs.remove(addr)
+
+        elif addr == from_addr:
+            from_contact = c
+
+    return contacts, to_contacts, from_contact
+
+
+def sanitize_params(params, all_args, required_args=None):
+    """Removes parameters that aren't valid.
+
+    Required Arguments:
+        params: dict - key/value pairs of arguments
+        all_args: list - list of strings, each string is a
+            valid parameter.
+
+    Optional Args:
+        required_arguments: list - ironically, required_arguments is an
+            optional argument here. a list of string, each string is a
+            required argument.
+
+    Returns:
+        dictionary of key, value pairs of valid parameters.
+    """
+    if required_args:
+        # check to be sure we have all the required params
+        missing_required_args = []
+        for required_arg in required_args:
+            param = params.get(required_arg)
+            if param == None:
+                missing_required_args.append(required_arg)
+
+        # yell if we're missing a required argument
+        if missing_required_args:
+            raise ArgumentError(
+                'Missing the following required arguments: %s' \
+                % ', '.join(missing_required_args)
+            )
+
+    # remove any arguments not recognized
+    cleaned_args = {}
+    for arg in all_args:
+        if arg in params:
+            cleaned_args[arg] = params[arg]
+            del params[arg]
+
+    # quietly yell in a non-breaking way if there's any unrecognized
+    # arguments left
+    if params:
+        logging.warning('Invalid arguments found: %s' % \
+            ', '.join(param for param in params))
+
+    return cleaned_args
