+++ dev-requirements.txt
@@ -1,0 +1,4 @@
+mock==1.3.0
+httpretty==0.8.14
+nose==1.3.7
+rednose==0.4.3
