+++ requirements.txt
@@ -1,0 +1,3 @@
+rauth==0.7.2
+requests==2.9.1
+six==1.10.0
