+++ tests/lib/v2_0/resources/test_contact.py
@@ -1,0 +1,124 @@
+import json
+import httpretty
+import unittest
+
+from contextio.contextio import ContextIO
+from contextio.lib.v2_0.resources.account import Account
+from contextio.lib.v2_0.resources.contact import Contact
+from contextio.lib.v2_0.resources.file import File
+from contextio.lib.v2_0.resources.message import Message
+from contextio.lib.v2_0.resources.thread import Thread
+
+class TestContact(unittest.TestCase):
+    def setUp(self):
+        self.contextio = ContextIO(consumer_key="foo", consumer_secret="bar")
+        self.account = Account(self.contextio, {"id": "fake_id"})
+
+        contact_definition = {
+            "email": "fake@email.com",
+            "count": 1,
+            "sent_count": 0,
+            "received_count": 1,
+            "sent_from_account_count": 0,
+            "thumbnail": "https://some.url",
+            "last_sent": None,
+            "last_received": 1455649370
+        }
+
+        self.contact = Contact(self.account, contact_definition)
+
+    def test_constructor_sets_attributes_on_object(self):
+        self.assertEqual("fake@email.com", self.contact.email)
+        self.assertEqual(1, self.contact.count)
+        self.assertEqual(0, self.contact.sent_count)
+        self.assertEqual(1, self.contact.received_count)
+        self.assertEqual(0, self.contact.sent_from_account_count)
+        self.assertEqual("https://some.url", self.contact.thumbnail)
+        self.assertEqual(None, self.contact.last_sent)
+        self.assertEqual(1455649370, self.contact.last_received)
+
+    def test_constructor_creates_emails_list_if_doesnt_exist(self):
+        self.assertEqual(["fake@email.com"], self.contact.emails)
+
+    @httpretty.activate
+    def test_get_updates_attributes_and_returns_True(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            "https://api.context.io/2.0/accounts/fake_id/contacts/fake@email.com/",
+            status=200,
+            body=json.dumps({
+                "email": "new@email.com",
+                "count": 5,
+                "sent_count": 200,
+                "received_count": 101,
+                "sent_from_account_count": 600,
+                "thumbnail": "https://some.url",
+                "last_sent": 2000000000,
+                "last_received": 2000000000
+            })
+        )
+
+        contact_updated = self.contact.get()
+
+        self.assertTrue(contact_updated)
+        self.assertEqual("new@email.com", self.contact.email)
+        self.assertEqual(5, self.contact.count)
+        self.assertEqual(200, self.contact.sent_count)
+        self.assertEqual(101, self.contact.received_count)
+        self.assertEqual(600, self.contact.sent_from_account_count)
+        self.assertEqual("https://some.url", self.contact.thumbnail)
+        self.assertEqual(2000000000, self.contact.last_sent)
+        self.assertEqual(2000000000, self.contact.last_received)
+
+
+    @httpretty.activate
+    def test_get_updates_email_from_list_of_emails(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            "https://api.context.io/2.0/accounts/fake_id/contacts/fake@email.com/",
+            status=200,
+            body=json.dumps({"emails": ["new@email.com"]}))
+
+        self.contact.get()
+
+        self.assertEqual("new@email.com", self.contact.email)
+
+    @httpretty.activate
+    def test_get_files_returns_list_of_files(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            "https://api.context.io/2.0/accounts/fake_id/contacts/fake@email.com/files",
+            status=200,
+            body=json.dumps([{"file_id": "foobar"}]))
+
+        contact_files = self.contact.get_files()
+
+        self.assertEqual(1, len(contact_files))
+        self.assertIsInstance(contact_files[0], File)
+
+    @httpretty.activate
+    def test_get_files_returns_list_of_messages(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            "https://api.context.io/2.0/accounts/fake_id/contacts/fake@email.com/messages",
+            status=200,
+            body=json.dumps([{"message_id": "foobar"}]))
+
+        contact_messages = self.contact.get_messages()
+
+        self.assertEqual(1, len(contact_messages))
+        self.assertIsInstance(contact_messages[0], Message)
+
+    @httpretty.activate
+    def test_get_files_returns_list_of_threads(self):
+        httpretty.register_uri(
+            httpretty.GET,
+            "https://api.context.io/2.0/accounts/fake_id/contacts/fake@email.com/threads",
+            status=200,
+            body=json.dumps(["https://thread.urls/thread"]))
+
+        contact_threads = self.contact.get_threads()
+
+        self.assertEqual(1, len(contact_threads))
+        self.assertIsInstance(contact_threads[0], Thread)
+
